Convert ECLDigitWaveformParameters from vector double to float array

+++ ecl/modules/eclWaveformFit/include/ECLWaveformFit.h
@@ -69,13 +69,13 @@ namespace Belle2 {
     { return "ECLDsps" ; }
 
   private:
-    std::vector<double> FitWithROOT(double, std::vector<double>, std::vector<double>, int);  /** Fit with ROOT::Fit function.*/
+    std::vector<double> FitWithROOT(double, std::vector<float>, std::vector<float>, int);  /** Fit with ROOT::Fit function.*/
     double m_CurrentPulseArray31[Belle2::ECL::EclConfiguration::m_nsmp];  /** Current waveform adc values.*/
     int m_FitType;  /**0 = photon + hadron, 1 = photon + diode*/
     double m_EnergyThreshold;  /**energy threshold to fit pulse offline*/
     double m_TriggerThreshold;  /**energy threshold for waveform trigger.*/
     std::vector<double> m_ADCtoEnergy;  /**calibration vector form adc to energy*/
-    std::vector< std::vector<double> > m_PhotonTemplates; /**photon templates*/
-    std::vector< std::vector<double> > m_SecondComponentTemplates; /**hadron or diode templates*/
+    std::vector< std::vector<float> > m_PhotonTemplates; /**photon templates*/
+    std::vector< std::vector<float> > m_SecondComponentTemplates; /**hadron or diode templates*/
   };
 } // end Belle2 namespace
